Se agrega validacion en clase Crypto y se organiza package de tests para el modelo
